Implement display of reboot times in graphs
If the sarfile is part of an sosreport, /var/log/messages* will be
parsed and reboot times will be noted and displayed in the graphs.

As the parsing of the correct dates in /var/log/messages is quite
fragile, this feature is off by default. Set SHOW_REBOOTS = True
in SAR.py to use it.