Fix a bug when a directory is passed as argument
Previously sarstats tried to look for sar files based on their
mtime ordering. As this is too fragile in sosreports, as often
the sar files' mtimes are all the same, we now actually parse
the first line of the sar file and interpret the date.
Only then we sort the sar files by date. This makes it
robust enough.